% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sback.R
\name{plot.sback}
\alias{plot.sback}
\title{Default sback plotting}
\usage{
\method{plot}{sback}(x, ask = TRUE, select = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{sback} as produced by \code{sback()}.}

\item{ask}{a logical value. If \code{TRUE}, the default, the user is asked for confirmation, before a new figure is drawn.}

\item{select}{Allows the plot for a single model term to be selected for printing. e.g. if you just want the plot for the second smooth term set select = 2.}

\item{\dots}{other graphics parameters to pass on to plotting commands.}
}
\description{
Takes a fitted \code{sback} object produced by \code{sback()} and plot the estimated partial functions on the scale of the linear predictor
}
