% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vcoef.R
\name{plot.vcoef}
\alias{plot.vcoef}
\title{Default vcoef plotting}
\usage{
\method{plot}{vcoef}(x, ask = TRUE, select = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{vcoef} as produced by \code{vcoef()}.}

\item{ask}{a logical value. If \code{TRUE}, the default, the user is asked for confirmation, before a new figure is drawn.}

\item{select}{Allows the plot for a single model term to be selected for printing. e.g. if you just want the plot for the second smooth term set select=2.}

\item{\dots}{other graphics parameters to pass on to plotting commands.}
}
\description{
Takes a fitted \code{vcoef} object produced by \code{vcoef()} and plot the estimated partial functions on the scale of the linear predictor.
}
