% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sback.R
\name{print.sback}
\alias{print.sback}
\title{Print a \code{sback} object.}
\usage{
\method{print}{sback}(x, ...)
}
\arguments{
\item{x}{an object of class \code{sback} as produced by \code{sback()}.}

\item{\dots}{other arguments, ignored.}
}
\description{
The default print method for a \code{sback} object.
}
