% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.vcoef.R
\name{print.vcoef}
\alias{print.vcoef}
\title{Print a \code{vcoef} object.}
\usage{
\method{print}{vcoef}(x, ...)
}
\arguments{
\item{x}{an object of class \code{vcoef} as produced by \code{vcoef()}.}

\item{\dots}{other arguments, ignored.}
}
\description{
The default print method for a \code{vcoef} object.
}
