% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sback.R
\name{sback}
\alias{sback}
\alias{sbackMain}
\title{Generalized additive and partially linear models.}
\usage{
sback(formula, data, weights = NULL, kbin = 15, family = 2,
  newdata = NULL)
}
\arguments{
\item{formula}{TODO}

\item{data}{TODO}

\item{weights}{TODO}

\item{kbin}{TODO}

\item{family}{TODO}

\item{newdata}{TODO}
}
\description{
Main function for fitting generalized additive and partially linear models by using smooth backfitting.
}
