% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sback.R
\name{summary.sback}
\alias{summary.sback}
\title{Summary for a \code{sback} fitted object.}
\usage{
\method{summary}{sback}(object, ...)
}
\arguments{
\item{object}{an object of class \code{sback} as produced by \code{sback()}.}

\item{\dots}{other arguments (not implemented).}
}
\description{
Takes a fitted \code{sback} object produced by \code{sback()} and produces various useful summaries from it.
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (object, ...)
{
  class(object) <- "summary.sback"
  object
}
}
