% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.vcoef.R
\name{summary.vcoef}
\alias{summary.vcoef}
\title{Summary for a \code{vcoef} fitted object.}
\usage{
\method{summary}{vcoef}(object, ...)
}
\arguments{
\item{object}{an object of class \code{vcoef} as produced by \code{vcoef()}.}

\item{\dots}{other arguments (not implemented).}
}
\description{
Takes a fitted \code{vcoef} object produced by \code{vcoef()} and produces various useful summaries from it.
}
