% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcoef.R
\name{vcoef}
\alias{vcoef}
\alias{vcoefMain}
\title{Generalized varying coefficient models.}
\usage{
vcoef(formula, data, weights = NULL, kbin = 15, family = 2,
  newdata = NULL)
}
\arguments{
\item{formula}{TODO}

\item{data}{TODO}

\item{weights}{TODO}

\item{kbin}{TODO}

\item{family}{TODO}

\item{newdata}{TODO}
}
\description{
Main function for fitting generalized varying coefficient models.
}
